feat: detect redirections in an existing tab
This works for regular web navigation, where clicking a link
actually causes a request to a server for the whole page.

If the app handles the redirect itself and only replaces the
text in the URL (avoiding a full page reload), it is not detected.

TODO: use the `tabs` api to detect SPA style "redirects"